+++ test/unit/morpho_t.js
@@ -11,7 +11,7 @@ describe("Segmentation", function(){
 
   it("sentence segmentation", function(){
     let text = "Oh! How can you? It is not fair. You have to do something.";
-    let res = ["Oh", "How can you", "It is not fair", "You have to do something"];
+    let res = ["Oh", "!", "How can you", "?", "It is not fair", ".", "You have to do something", "."];
     expect(morpho.splitToSentences(text)).to.eql(res);
   });